refactor(patrol): extract shared Hive::OutputPulse for idle deadlines
The mutex-guarded monotonic output pulse now lives in
lib/hive/output_pulse.rb so the review CI runner (#1150) reuses one
thread-coordination primitive instead of a private copy.
